+++ packages/ragbits-document-search/src/ragbits/document_search/_main.py
@@ -14,7 +14,7 @@ from ragbits.core.llms.factory import get_preferred_llm
 from ragbits.core.utils._pyproject import get_config_from_yaml
 from ragbits.core.utils.config_handling import (
     NoPreferredConfigError,
-    ObjectContructionConfig,
+    ObjectConstructionConfig,
     WithConstructionConfig,
     import_by_path,
 )
@@ -55,12 +55,12 @@ class DocumentSearchConfig(BaseModel):
     Schema for the dict taken by DocumentSearch.from_config method.
     """
 
-    vector_store: ObjectContructionConfig
-    rephraser: ObjectContructionConfig = ObjectContructionConfig(type="NoopQueryRephraser")
-    reranker: ObjectContructionConfig = ObjectContructionConfig(type="NoopReranker")
-    ingest_strategy: ObjectContructionConfig = ObjectContructionConfig(type="SequentialIngestStrategy")
-    providers: dict[str, ObjectContructionConfig] = {}
-    intermediate_element_handlers: dict[str, ObjectContructionConfig] = {}
+    vector_store: ObjectConstructionConfig
+    rephraser: ObjectConstructionConfig = ObjectConstructionConfig(type="NoopQueryRephraser")
+    reranker: ObjectConstructionConfig = ObjectConstructionConfig(type="NoopReranker")
+    ingest_strategy: ObjectConstructionConfig = ObjectConstructionConfig(type="SequentialIngestStrategy")
+    providers: dict[str, ObjectConstructionConfig] = {}
+    intermediate_element_handlers: dict[str, ObjectConstructionConfig] = {}
 
 
 class DocumentSearch(WithConstructionConfig):
@@ -153,7 +153,7 @@ class DocumentSearch(WithConstructionConfig):
         yaml_path_override: Path | None = None,
     ) -> Self:
         """
-        Tries to create an instance by looking at project's component prefferences, either from YAML
+        Tries to create an instance by looking at project's component preferences, either from YAML
         or from the factory. Takes optional overrides for both, which takes a higher precedence.
 
         Args:
@@ -168,14 +168,14 @@ class DocumentSearch(WithConstructionConfig):
             InvalidConfigError: If the default factory or configuration can't be found.
         """
         if yaml_path_override:
-            preferrences = get_config_from_yaml(yaml_path_override)
+            preferences = get_config_from_yaml(yaml_path_override)
 
             # Look for explicit document search configuration
-            if type_config := preferrences.get(cls.configuration_key):
-                return cls.subclass_from_config(ObjectContructionConfig.model_validate(type_config))
+            if type_config := preferences.get(cls.configuration_key):
+                return cls.subclass_from_config(ObjectConstructionConfig.model_validate(type_config))
 
-            # Instantate the class with the preferred configuration for each component
-            return cls.from_config(preferrences)
+            # Instantiate the class with the preferred configuration for each component
+            return cls.from_config(preferences)
 
         if factory_path_override:
             return cls.subclass_from_factory(factory_path_override)
@@ -186,9 +186,9 @@ class DocumentSearch(WithConstructionConfig):
         if config.component_preference_config_path is not None:
             # Look for explicit document search configuration
             if preferred_config := config.preferred_instances_config.get(cls.configuration_key):
-                return cls.subclass_from_config(ObjectContructionConfig.model_validate(preferred_config))
+                return cls.subclass_from_config(ObjectConstructionConfig.model_validate(preferred_config))
 
-            # Instantate the class with the prefereed configuration for each component
+            # Instantiate the class with the preferred configuration for each component
             return cls.from_config(config.preferred_instances_config)
 
         raise NoPreferredConfigError(f"Could not find preferred factory or configuration for {cls.configuration_key}")