+++ packages/ragbits-document-search/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- correcting typo in class name ObjectConstructionConfig and other typo corrections
+
 ## 0.10.0 (2025-03-17)
 
 ### Changed
@@ -9,12 +11,11 @@
 - ragbits-core updated to version v0.10.0
 
 - BREAKING CHANGE: Processing strategies refactored to ingest strategies (#394)
-- Compability with the new Vector Store interface from ragbits-core (#288)
+- Compatibility with the new Vector Store interface from ragbits-core (#288)
 - Fix docstring formatting to resolve Griffe warnings
 - Introduce intermediate image elements (#139)
 - Add HTTP source type, which downloads a file from the provided URL (#397)
-
- - added traceable
+- added traceable
 
 ## 0.9.0 (2025-02-25)